Improved coding segments for av
Audio/video segments can be coded, deleted, memoed. Segments are shown as color-coded strips in upper half of the gui. Mouse hover shows the code name, time stamps and any memo.
Right click on a code strip gives a menu to edit the memo, or to delete the coded segment.
Coded strips are shown so that they do not overlap.