Added backup options
Backup options include: no backup on loading a project and backups that exclude audio /video files. A/V files are large and so backups can be slow and use a lot of space. Also made some minor changes to the Qualcoder_settings.txt file. Changes to this include stating the setting name, a colon, and a value. Previously variable names were not shown.