Improve setup.py and switch to relative imports
With the old setup.py I was not able to install the program so I tried
to fix the setup.py.

Now it creates a correct wheel, and creates a "qualcoder" executable
that will be installed into the /bin/ folder that setup-tools uses to
install software.

Using this install method the unqualified import didn‘t work, so I made
them relative.

I can not test this on windows, but I assume there has to be a way to
use setup.py to create an executable similar to the linux "qualcoder"
executable.