Fix import issues discussed in issue #500
qualcoder.py was renamed to __main__.py because Python will throw
incredibly confusing errors if a you have a qualcoder.py _file_ with the same name
as a _package_ (qualcoder), because it will try to import the file instead of the
package, if the the former comes first in the PYTHON_PATH.

This also allows running the module directly, i.e. `python -m
qualcoder` will now use python to resolve the the path of the module
qualcoder, and then basically run `python path/to/qualcoder/__main__.py`

The imports in this __main__.py are also changed to absolute imports,
because if the file is run directly, it isn't imported as a module, and
there is no module the imports could be relative to. The imports in the
other files can be left relative, because those files are imported
themselves, so a module to be relative to exists at that time.