add initial flatpak manifest
Initial working prototype
Adapted from
- https://docs.flatpak.org/en/latest/python.html
- https://develop.kde.org/docs/getting-started/python/python-flatpak/
- https://github.com/flathub/org.thonny.Thonny/blob/master/org.thonny.Thonny.yaml

Flatpak requires all dependencies to be specified and arbitrary downloads during the build are not possible.
Thus all python dependencies are specified in flatpak-pypi-dependencies.json which was generated
using req2flatpak https://github.com/johannesjh/req2flatpak
That tool requires a requirements.txt with exact version numbers, which is in requirements-flatpak.txt and
was created with uv. The pyproject.toml for uv was created with `uv add -r requirements.txt`.
The pyproject.toml and uv.lock are not included in this commit, as uv is a separate dependency management
and build system.
The flatpak-pip-generator from the flatpak-builder-tools was considered as well for generating these
dependency definitions directly from the requirements.txt, but then all dependencies would be build from
source, which did not succeed for me.
- https://github.com/flatpak/flatpak-builder-tools/tree/master/pip

The resulting package is not optimized and thus includes all full dependencies.
Build directories amass 8 GB and the compressed resulting packages is just under 3 GB.

It is also missing a desktop file and appstream metadata.